+++ example/lib/main.dart
@@ -30,9 +30,11 @@ class _MyAppState extends State<MyApp> {
       home: Theme(
         data: ThemeData(
           primaryColor: const Color(0xFF2C1C6B),
-          colorScheme: ColorScheme.light().copyWith(secondary: const Color(0xFF200681)),
+          colorScheme:
+              ColorScheme.light().copyWith(secondary: const Color(0xFF200681)),
           cardColor: const Color(0xFFF8F9FC),
-          textTheme: const TextTheme(bodyText2: TextStyle(fontSize: 20)),
+          textTheme: const TextTheme(
+              bodyLarge: TextStyle(fontSize: 20, color: Color(0xFF2C1C6B))),
         ),
         child: Scaffold(
           appBar: AppBar(
@@ -56,6 +58,7 @@ class _MyAppState extends State<MyApp> {
                           maxLines: 10,
                           actions: MarkdownType.values,
                           controller: controller,
+                          textStyle: TextStyle(fontSize: 16),
                         ),
                         TextButton(
                           onPressed: () {