+++ lib/format_markdown.dart
@@ -1,3 +1,4 @@
+import 'dart:math';
 import 'package:flutter/material.dart';
 
 /// Use this class for converting String to [ResultMarkdown]
@@ -5,10 +6,17 @@ class FormatMarkdown {
   /// Convert [data] part into [ResultMarkdown] from [type].
   /// Use [fromIndex] and [toIndex] for converting part of [data]
   /// [titleSize] is used for markdown titles
+  /// [link] is used for link conversion type
   static ResultMarkdown convertToMarkdown(MarkdownType type, String data, int fromIndex, int toIndex,
-      {int titleSize = 1}) {
+      {int titleSize = 1, String? link, String selectedText = ''}) {
     late String changedData;
     late int replaceCursorIndex;
+
+    final lesserIndex = min(fromIndex, toIndex);
+    final greaterIndex = max(fromIndex, toIndex);
+
+    fromIndex = lesserIndex;
+    toIndex = greaterIndex;
 
     switch (type) {
       case MarkdownType.bold:
@@ -24,8 +32,8 @@ class FormatMarkdown {
         replaceCursorIndex = 2;
         break;
       case MarkdownType.link:
-        changedData = '[${data.substring(fromIndex, toIndex)}](${data.substring(fromIndex, toIndex)})';
-        replaceCursorIndex = 3;
+        changedData = '[$selectedText](${link ?? selectedText})';
+        replaceCursorIndex = 0;
         break;
       case MarkdownType.title:
         changedData = "${"#" * titleSize} ${data.substring(fromIndex, toIndex)}";