+++ test/markdown_text_input_test.dart
@@ -1,11 +1,24 @@
 import 'package:flutter/material.dart';
-import 'package:flutter/widgets.dart';
 import 'package:flutter_test/flutter_test.dart';
+import 'package:markdown_editable_textinput/format_markdown.dart';
 import 'package:markdown_editable_textinput/markdown_text_input.dart';
 
 void main() {
+  Widget component(String initialValue, ValueChanged<String> updateValue) =>
+      MaterialApp(
+        home: Scaffold(
+          body: MarkdownTextInput(
+            updateValue,
+            initialValue,
+            actions: MarkdownType.values,
+            insertLinksByDialog: false,
+          ),
+        ),
+      );
+
   testWidgets('MarkdownTextInput has all buttons', (WidgetTester tester) async {
-    await tester.pumpWidget(MaterialApp(home: Scaffold(body: MarkdownTextInput(print, 'initial value'))));
+    await tester.pumpWidget(component('initial value', (_) => null));
+
     final boldKey = const Key('bold_button');
     final italicKey = const Key('italic_button');
     final strikethroughKey = const Key('strikethrough_button');
@@ -39,16 +52,15 @@ void main() {
     expect(find.byKey(imageKey), findsOneWidget);
   });
 
-  testWidgets('MarkdownTextInput make bold from selection', (WidgetTester tester) async {
+  testWidgets('MarkdownTextInput make bold from selection',
+      (WidgetTester tester) async {
     var initialValue = 'initial value';
-    await tester.pumpWidget(MaterialApp(
-        home: Scaffold(
-            body: MarkdownTextInput((String value) {
-      initialValue = value;
-    }, initialValue))));
+    await tester.pumpWidget(
+        component(initialValue, (String value) => initialValue = value));
 
     final formfield = tester.widget<EditableText>(find.text(initialValue));
-    formfield.controller.selection = TextSelection(baseOffset: 0, extentOffset: initialValue.length);
+    formfield.controller.selection =
+        TextSelection(baseOffset: 0, extentOffset: initialValue.length);
 
     final boldKey = const Key('bold_button');
     await tester.tap(find.byKey(boldKey));
@@ -56,16 +68,15 @@ void main() {
     expect(initialValue, '**initial value**');
   });
 
-  testWidgets('MarkdownTextInput make italic from selection', (WidgetTester tester) async {
+  testWidgets('MarkdownTextInput make italic from selection',
+      (WidgetTester tester) async {
     var initialValue = 'initial value';
-    await tester.pumpWidget(MaterialApp(
-        home: Scaffold(
-            body: MarkdownTextInput((String value) {
-      initialValue = value;
-    }, initialValue))));
+    await tester.pumpWidget(
+        component(initialValue, (String value) => initialValue = value));
 
     final formfield = tester.widget<EditableText>(find.text(initialValue));
-    formfield.controller.selection = TextSelection(baseOffset: 0, extentOffset: initialValue.length);
+    formfield.controller.selection =
+        TextSelection(baseOffset: 0, extentOffset: initialValue.length);
 
     final boldKey = const Key('italic_button');
     await tester.tap(find.byKey(boldKey));
@@ -73,16 +84,15 @@ void main() {
     expect(initialValue, '_initial value_');
   });
 
-  testWidgets('MarkdownTextInput make strikethrough from selection', (WidgetTester tester) async {
+  testWidgets('MarkdownTextInput make strikethrough from selection',
+      (WidgetTester tester) async {
     var initialValue = 'initial value';
-    await tester.pumpWidget(MaterialApp(
-        home: Scaffold(
-            body: MarkdownTextInput((String value) {
-      initialValue = value;
-    }, initialValue))));
+    await tester.pumpWidget(
+        component(initialValue, (String value) => initialValue = value));
 
     final formfield = tester.widget<EditableText>(find.text(initialValue));
-    formfield.controller.selection = TextSelection(baseOffset: 0, extentOffset: initialValue.length);
+    formfield.controller.selection =
+        TextSelection(baseOffset: 0, extentOffset: initialValue.length);
 
     final boldKey = const Key('strikethrough_button');
     await tester.tap(find.byKey(boldKey));
@@ -90,16 +100,15 @@ void main() {
     expect(initialValue, '~~initial value~~');
   });
 
-  testWidgets('MarkdownTextInput make code from selection', (WidgetTester tester) async {
+  testWidgets('MarkdownTextInput make code from selection',
+      (WidgetTester tester) async {
     var initialValue = 'initial value';
-    await tester.pumpWidget(MaterialApp(
-        home: Scaffold(
-            body: MarkdownTextInput((String value) {
-      initialValue = value;
-    }, initialValue))));
+    await tester.pumpWidget(
+        component(initialValue, (String value) => initialValue = value));
 
     final formfield = tester.widget<EditableText>(find.text(initialValue));
-    formfield.controller.selection = TextSelection(baseOffset: 0, extentOffset: initialValue.length);
+    formfield.controller.selection =
+        TextSelection(baseOffset: 0, extentOffset: initialValue.length);
 
     final boldKey = const Key('code_button');
     await tester.tap(find.byKey(boldKey));
@@ -107,16 +116,15 @@ void main() {
     expect(initialValue, '```initial value```');
   });
 
-  testWidgets('MarkdownTextInput make link from selection', (WidgetTester tester) async {
+  testWidgets('MarkdownTextInput make link from selection',
+      (WidgetTester tester) async {
     var initialValue = 'initial value';
-    await tester.pumpWidget(MaterialApp(
-        home: Scaffold(
-            body: MarkdownTextInput((String value) {
-      initialValue = value;
-    }, initialValue))));
+    await tester.pumpWidget(
+        component(initialValue, (String value) => initialValue = value));
 
     final formfield = tester.widget<EditableText>(find.text(initialValue));
-    formfield.controller.selection = TextSelection(baseOffset: 0, extentOffset: initialValue.length);
+    formfield.controller.selection =
+        TextSelection(baseOffset: 0, extentOffset: initialValue.length);
 
     final boldKey = const Key('link_button');
     await tester.tap(find.byKey(boldKey));
@@ -124,16 +132,15 @@ void main() {
     expect(initialValue, '[initial value](initial value)');
   });
 
-  testWidgets('MarkdownTextInput make list from selection', (WidgetTester tester) async {
+  testWidgets('MarkdownTextInput make list from selection',
+      (WidgetTester tester) async {
     var initialValue = 'initial\nvalue';
-    await tester.pumpWidget(MaterialApp(
-        home: Scaffold(
-            body: MarkdownTextInput((String value) {
-      initialValue = value;
-    }, initialValue))));
+    await tester.pumpWidget(
+        component(initialValue, (String value) => initialValue = value));
 
     final formfield = tester.widget<EditableText>(find.text(initialValue));
-    formfield.controller.selection = TextSelection(baseOffset: 0, extentOffset: initialValue.length);
+    formfield.controller.selection =
+        TextSelection(baseOffset: 0, extentOffset: initialValue.length);
 
     final boldKey = const Key('list_button');
     await tester.tap(find.byKey(boldKey));
@@ -141,16 +148,15 @@ void main() {
     expect(initialValue, '* initial\n* value');
   });
 
-  testWidgets('MarkdownTextInput make blockquote from selection', (WidgetTester tester) async {
+  testWidgets('MarkdownTextInput make blockquote from selection',
+      (WidgetTester tester) async {
     var initialValue = 'initial\nvalue';
-    await tester.pumpWidget(MaterialApp(
-        home: Scaffold(
-            body: MarkdownTextInput((String value) {
-      initialValue = value;
-    }, initialValue))));
+    await tester.pumpWidget(
+        component(initialValue, (String value) => initialValue = value));
 
     final formfield = tester.widget<EditableText>(find.text(initialValue));
-    formfield.controller.selection = TextSelection(baseOffset: 0, extentOffset: initialValue.length);
+    formfield.controller.selection =
+        TextSelection(baseOffset: 0, extentOffset: initialValue.length);
 
     final boldKey = const Key('quote_button');
     await tester.tap(find.byKey(boldKey));
@@ -158,16 +164,15 @@ void main() {
     expect(initialValue, '> initial\n> value');
   });
 
-  testWidgets('MarkdownTextInput make separator from selection', (WidgetTester tester) async {
+  testWidgets('MarkdownTextInput make separator from selection',
+      (WidgetTester tester) async {
     var initialValue = 'initial value';
-    await tester.pumpWidget(MaterialApp(
-        home: Scaffold(
-            body: MarkdownTextInput((String value) {
-      initialValue = value;
-    }, initialValue))));
+    await tester.pumpWidget(
+        component(initialValue, (String value) => initialValue = value));
 
     final formfield = tester.widget<EditableText>(find.text(initialValue));
-    formfield.controller.selection = TextSelection(baseOffset: 0, extentOffset: initialValue.length);
+    formfield.controller.selection =
+        TextSelection(baseOffset: 0, extentOffset: initialValue.length);
 
     final boldKey = const Key('separator_button');
     await tester.tap(find.byKey(boldKey));
@@ -175,16 +180,15 @@ void main() {
     expect(initialValue, '\n------\ninitial value');
   });
 
-  testWidgets('MarkdownTextInput make image link from selection', (WidgetTester tester) async {
+  testWidgets('MarkdownTextInput make image link from selection',
+      (WidgetTester tester) async {
     var initialValue = 'initial value';
-    await tester.pumpWidget(MaterialApp(
-        home: Scaffold(
-            body: MarkdownTextInput((String value) {
-      initialValue = value;
-    }, initialValue))));
+    await tester.pumpWidget(
+        component(initialValue, (String value) => initialValue = value));
 
     final formfield = tester.widget<EditableText>(find.text(initialValue));
-    formfield.controller.selection = TextSelection(baseOffset: 0, extentOffset: initialValue.length);
+    formfield.controller.selection =
+        TextSelection(baseOffset: 0, extentOffset: initialValue.length);
 
     final boldKey = const Key('image_button');
     await tester.tap(find.byKey(boldKey));